WIP: switch the toptalk chart y-axis to sqrt scale
One of the first things people say when they see the jt demo is:
"you should use a log scale on the y-axis". However, log(0) is really
inconvenient (and non-linear scales have other drawbacks).

This change changes the toptalk chart to a square root scale, but is
really work-in-progress. A selector between linear and sqrt will follow.